requirements.txt
transformers
Pillow
nltk
networkx
matplotlib
torch
requests
TTS
Flask
werkzeug
